third_party/blink/web_tests/external/wpt/shadow-dom/leaktests/html-collection-expected.txt

This is a testharness.js-based test.
[FAIL] document.applets should not contain any nodes
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
